collection_validators.py created from prod_validators.py
